+++ tt-media-server/cpp_server/include/ipc/in_memory/in_memory_memory_queue.hpp
@@ -3,7 +3,8 @@
 
 #pragma once
 
-#include "ipc/in_memory/detail/concurrent_queue.hpp"
+
+#include "utils/concurrent_queue.hpp"
 #include "ipc/interface/memory_queue.hpp"
 
 namespace tt::ipc::in_memory {
@@ -19,7 +20,7 @@ class MemoryRequestQueue : public tt::ipc::IMemoryRequestQueue {
   }
 
  private:
-  tt::ipc::in_memory::detail::ConcurrentQueue<tt::domain::ManageMemoryTask>
+  tt::utils::ConcurrentQueue<tt::domain::ManageMemoryTask>
       queue;
 };
 
@@ -34,7 +35,7 @@ class MemoryResultQueue : public tt::ipc::IMemoryResultQueue {
   }
 
  private:
-  tt::ipc::in_memory::detail::ConcurrentQueue<tt::domain::ManageMemoryResult>
+  tt::utils:ConcurrentQueue<tt::domain::ManageMemoryResult>
       queue;
 };
