+++ tt-media-server/cpp_server/include/ipc/in_memory/in_memory_result_queue.hpp
@@ -5,8 +5,9 @@
 
 #include <chrono>
 
-#include "ipc/in_memory/detail/concurrent_queue.hpp"
+
 #include "ipc/interface/result_queue.hpp"
+#include "utils/concurrent_queue.hpp"
 
 namespace tt::ipc::in_memory {
 
@@ -34,7 +35,7 @@ class ResultQueue : public tt::ipc::IResultQueue {
   }
 
  private:
-  tt::ipc::in_memory::detail::ConcurrentQueue<tt::ipc::SharedToken> queue;
+  tt::utils::ConcurrentQueue<tt::ipc::SharedToken> queue;
 };
 
 }  // namespace tt::ipc::in_memory
