+++ tt-media-server/cpp_server/include/utils/concurrent_queue.hpp
@@ -7,6 +7,8 @@
 #include <cstddef>
 #include <mutex>
 #include <vector>
+#include <condition_variable>
+#include <deque>
 
 #include "profiling/tracy.hpp"
 
@@ -144,6 +146,87 @@ class LockFreeSPSCQueue {
   std::vector<T> buffer;
   alignas(detail::CACHE_LINE_SIZE) std::atomic<size_t> head{0};
   alignas(detail::CACHE_LINE_SIZE) std::atomic<size_t> tail{0};
+};
+
+template <typename T>
+class BlockingQueue {
+ public:
+  void push(T item) {
+    {
+      std::lock_guard<std::mutex> lock(mu);
+      queue.push_back(std::move(item));
+    }
+    cv.notify_one();
+  }
+
+  bool tryPop(T& out) {
+    std::lock_guard<std::mutex> lock(mu);
+    if (queue.empty()) {
+      return false;
+    }
+    out = std::move(queue.front());
+    queue.pop_front();
+    return true;
+  }
+
+  bool waitPop(T& out) {
+    std::unique_lock<std::mutex> lock(mu);
+    cv.wait(lock, [&] { return shuttingDown || !queue.empty(); });
+    if (queue.empty()) {
+      return false;
+    }
+    out = std::move(queue.front());
+    queue.pop_front();
+    return true;
+  }
+
+  template <typename Rep, typename Period>
+  bool waitPopFor(T& out, std::chrono::duration<Rep, Period> timeout) {
+    std::unique_lock<std::mutex> lock(mu);
+    if (!cv.wait_for(lock, timeout,
+                     [&] { return shuttingDown || !queue.empty(); })) {
+      return false;
+    }
+    if (queue.empty()) {
+      return false;
+    }
+    out = std::move(queue.front());
+    queue.pop_front();
+    return true;
+  }
+
+  template <typename Container>
+  void drainTo(Container& out) {
+    std::lock_guard<std::mutex> lock(mu);
+    while (!queue.empty()) {
+      out.push_back(std::move(queue.front()));
+      queue.pop_front();
+    }
+  }
+
+  void clear() {
+    std::lock_guard<std::mutex> lock(mu);
+    queue.clear();
+  }
+
+  void shutdown() {
+    {
+      std::lock_guard<std::mutex> lock(mu);
+      shuttingDown = true;
+    }
+    cv.notify_all();
+  }
+
+  bool empty() const {
+    std::lock_guard<std::mutex> lock(mu);
+    return queue.empty();
+  }
+
+ private:
+  mutable std::mutex mu;
+  std::condition_variable cv;
+  std::deque<T> queue;
+  bool shuttingDown = false;
 };
 
 }  // namespace tt::utils
